Fix missing stock initialiser
